Added a new Admin Service System and Human Interface
Added the first menu, which asks for admin authetification. After that, the banker has a small menu which requires a certain password for each part, besides the customers one, where he can go without any other password.

Also added the human interface, which is the base for Customer and for Banker. Oh, and also a new Banker class.